requirements.txt
torch==2.7.1
torchvision==0.22.1
pillow==11.2.1
streamlit==1.45.1
